310 Upgrade Terminal
This push is a change to the Upgrade Terminal so it accesses our 'Widget Terminal 1' instead of the current 'Widget Terminal' that was previously there. Don't know if you want these changes, Nick. It prompted us to push this after I saved and closed Unreal (Which I should have done before first commit)

- Jordan Wynalda/Nate Connette